+++ content/themes/dude/template-parts/blocks/keynumbers-content.php
@@ -4,8 +4,8 @@
  *
  * Block that has keynumbers.
  *
- * @Author:		Roni Laukkarinen
- * @Date:   		2022-02-10 12:28:36
+ * @Author:     Roni Laukkarinen
+ * @Date:           2022-02-10 12:28:36
  * @Last Modified by:   Roni Laukkarinen
  * @Last Modified time: 2022-09-08 18:05:14
  *
@@ -36,16 +36,16 @@ if ( ! isset( $args ) ) {
         <?php foreach ( $keynumbers as $keynumber ) : ?>
           <div class="number">
             <h2 class="title">
-              <?php echo esc_html( $keynumber['title'] ) ?>
+          <?php echo esc_html( $keynumber['title'] ) ?>
             </h2>
 
             <div class="number-and-description">
               <p class="number">
-                <?php echo esc_html( $keynumber['number'] ) ?>
+          <?php echo esc_html( $keynumber['number'] ) ?>
               </p>
 
               <p class="description">
-                <?php echo esc_html( $keynumber['description'] ) ?>
+          <?php echo esc_html( $keynumber['description'] ) ?>
               </p>
             </div>
           </div>
@@ -58,6 +58,17 @@ if ( ! isset( $args ) ) {
 
           <?php echo wp_kses_post( $additional_info_content ) ?>
         </div>
+
+        <?php
+        // Show budget class for references
+        $budget_terms = wp_get_post_terms( get_the_ID(), 'reference-budget' );
+        if ( ! empty( $budget_terms ) && ! is_wp_error( $budget_terms ) ) :
+          ?>
+          <div class="budget-information">
+            <h2 class="additional-info-title">Budjettiluokka</h2>
+            <p class="budget-price"><?php echo esc_html( $budget_terms[0]->name ); ?></p>
+          </div>
+        <?php endif; ?>
       </div>
 
       <div class="col col-content">