+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 7.0.15: 2026-02-19
 
+* Add `reference-budget` taxonomy (Budjettiluokka) for references with filter support on archive page, Ref: DEV-765
+* Add budjettiluokka display in keynumbers-content additional-information section, Ref: DEV-765
 * Fix checkamrk bullet alignment, Ref: DEV-701
 
 ### 7.0.14: 2026-02-03